CreditCardScript
This script handles the interaction of the credit card number and name. During its initial creation, it checks the saved player preferences whether or not the name and number have already been saved. If the saved registry values are not null, then they are inserted into the text box. Its only method is called when the Add Card button is clicked. When clicked, it checks whether or not the two input fields are empty and if not, take the values and save them into the player preferences as strings.