HealthBar
This script handles the health bar for the enemies. Each enemy has a health bar that are child components, and are given positions to always be in the same position reference its parent enemy object. The health bar has a rectangular image that is scaled in relation to the current health and maximum health of the enemy.