MusicScript
This script is attached to the two music switch buttons in the options menu. If the music is currently on and the player chooses to click the off button then the music stops and the player preferences is saved with a new value indicating that no music will be played in the game. And vice versa, if the music is currently off.