ProjectileScript
This script is added to the bullets and rockets that the units shoot at their targeted enemies. The projectiles have a damage that is set based on already preset damage values for each unit and unit level. The projectile moves towards the enemy target and if it colliders, it damages the enemy based on the damage value. If the projectile kills the enemy target, it checks to see the type of enemy in order to determine a value of quarters to reward the player.