SFXScript
This script is attached to the two SFX switch buttons in the options menu. If the SFX is currently on and the player chooses to click the off button then the SFX stops and the player preferences is saved with a new value indicating that no SFX will be played in the game. And vice versa, if the SFX is currently off.