UnitShoot
This script is given to the units that are created which handle their ability to shoot the enemies. The unit targets the enemy in its collider range that is closest to the objective based on its distance. When allowed to shoot based on its firing rate, the unit instantiates a new projectile based of its level projectile prefab that is given a target position to move towards.